# cspt-metric-service/oas/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/performance-metrics": {
      "post": {
        "tags": [
          "api-controller"
        ],
        "operationId": "savePerformanceMetrics",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MetricDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/network-data": {
      "post": {
        "tags": [
          "api-controller"
        ],
        "operationId": "saveNetworkData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NetworkDataDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MetricDto": {
        "required": [
          "document",
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "document": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "NetworkDataDto": {
        "required": [
          "responseTime",
          "status",
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 100,
            "type": "integer",
            "format": "int32"
          },
          "responseTime": {
            "type": "number",
            "format": "float"
          }
        }
      }
    }
  }
}